refactor(gpu): split types.rs into registers.rs and primitives.rs
Split the large GPU types.rs file (997 lines) into focused modules:
- registers.rs: GPU register-related types (GPUStatus, DrawMode, etc.)
- primitives.rs: Rendering primitive types (Color, Vertex, BlendMode, etc.)

Updated all internal imports across 11 files in the GPU module.
No breaking changes to public API via re-exports.

Closes #149
